# user_mgmt/swagger_server/models/assign_context_permission_body.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class AssignContextPermissionBody(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self,
                 permission_id: str=None,
                 resource_id: str=None):  # noqa: E501
        """AssignContextPermissionBody - a model defined in Swagger

        :param permission_id: The permission_id of this AssignContextPermissionBody.  # noqa: E501
        :type permission_id: str
        :param resource_id: The resource_id of this AssignContextPermissionBody.  # noqa: E501
        :type resource_id: str
        """
        self.swagger_types = {'permission_id': str, 'resource_id': str}

        self.attribute_map = {
            'permission_id': 'permissionId',
            'resource_id': 'resourceId'
        }
        self._permission_id = permission_id
        self._resource_id = resource_id

    @classmethod
    def from_dict(cls, dikt) -> 'AssignContextPermissionBody':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The AssignContextPermissionBody of this AssignContextPermissionBody.  # noqa: E501
        :rtype: AssignContextPermissionBody
        """
        return util.deserialize_model(dikt, cls)

    @property
    def permission_id(self) -> str:
        """Gets the permission_id of this AssignContextPermissionBody.


        :return: The permission_id of this AssignContextPermissionBody.
        :rtype: str
        """
        return self._permission_id

    @permission_id.setter
    def permission_id(self, permission_id: str):
        """Sets the permission_id of this AssignContextPermissionBody.


        :param permission_id: The permission_id of this AssignContextPermissionBody.
        :type permission_id: str
        """

        self._permission_id = permission_id

    @property
    def resource_id(self) -> str:
        """Gets the resource_id of this AssignContextPermissionBody.


        :return: The resource_id of this AssignContextPermissionBody.
        :rtype: str
        """
        return self._resource_id

    @resource_id.setter
    def resource_id(self, resource_id: str):
        """Sets the resource_id of this AssignContextPermissionBody.


        :param resource_id: The resource_id of this AssignContextPermissionBody.
        :type resource_id: str
        """
        if resource_id is None:
            raise ValueError(
                "Invalid value for `resource_id`, must not be `None`")  # noqa: E501

        self._resource_id = resource_id